manager/requirements.txt
# Web Framework
Flask>=3.1.0
Flask-Security-Too==5.7.1
Flask-CORS==4.0.0
flask-sqlalchemy==3.1.1

# Database - PyDAL (mandatory for operations) and SQLAlchemy (init only)
pydal==20250215.2
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Data Validation
Pydantic==2.5.3

# gRPC Support
grpcio==1.60.1
grpcio-tools==1.60.1

# Caching and Async
redis==5.0.1
aioredis==2.0.1
aiohttp==3.9.1
httpx==0.25.2

# Cloud Providers
boto3>=1.26.0
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.15.0
azure-mgmt-sql>=3.0.0
azure-mgmt-redis>=14.0.0
azure-storage-blob>=12.15.0

# Kubernetes
kubernetes==29.0.0

# Security
cryptography==41.0.7
PyJWT==2.8.0
bcrypt==4.1.2

# WSGI Server
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1

# Monitoring
prometheus-client==0.19.0

# Additional utilities
requests==2.32.4
PyYAML==6.0.1
lxml==5.1.0
python-dotenv==1.0.0